hw23 fix  404

+++ src/quiz/useCase/DeleteQuestionHandler.ts
@@ -1,5 +1,5 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
-import { Inject } from "@nestjs/common";
+import { Inject, UnauthorizedException } from "@nestjs/common";
 import { IQuestionsRepo, IQuestionsRepoToken } from "../DAL/IQuestionsRepo";
 import { QuestionEntity } from "../DAL/QuestionEntity";
 
@@ -18,6 +18,8 @@ export class DeleteQuestionHandler
 
   async execute(command: DeleteQuestionCommand): Promise<void> {
     const { questionId } = command;
+    const question = await this.questionsRepo.getQuestionById(questionId);
+    if (!question) throw new UnauthorizedException("net takogo id");
     await this.questionsRepo.deleteQuestion(questionId);
     return;
   }
